搜集数据，全程用ruckig规划的误差数据

+++ Module/robot/include/MhIndustrialSCARA.h
@@ -88,7 +88,7 @@ public:
 public:
     MhRobotDataText MhRobotText;//关于文本记录的类
     bool judge=false;
-    int count=0;//记录发送的指令次数
+    double count=0;//记录发送的指令次数
 //-------------------------------------------------------视觉伺服误差相关变量
     double threshold_t;
     double threshold_tu;
